pyproject.toml
[tool.poetry]
name = "databack"
version = "0.1.0"
description = "Backup you data from MySQL/PostgreSQL etc. to any other storages"
authors = ["long2ice <long2ice@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { version = "*", extras = ["all"] }
tortoise-orm = { version = "*", extras = ["all"] }
sentry-sdk = { version = "*", extras = ["fastapi"] }
loguru = "*"
aerich = "*"
asyncmy = "*"
asyncpg = "*"
rearq = { git = "https://github.com/long2ice/rearq.git", branch = "dev" }
aioshutil = "*"
aioboto3 = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
mypy = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-custom-exit-code = "*"
ruff = "*"
types-aiofiles = "*"


[tool.aerich]
tortoise_orm = "databack.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310']

[tool.pytest.ini_options]
asyncio_mode = 'auto'

[tool.mypy]
ignore_missing_imports = true
pretty = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
